+++ examples/terraform/iam/loyaleasy-min/outputs.tf
@@ -1,0 +1,7 @@
+output "rds_proxy_role_arn" {
+  value = aws_iam_role.proxy.arn
+}
+
+output "rds_proxy_policy_arn" {
+  value = aws_iam_policy.rds_proxy.arn
+}